Fixed file 6

+++ 0x00-hello_world/6-size.c
@@ -6,10 +6,10 @@
  */
 int main(void)
 {
-printf("Size of char: %d bytes\n", (int)sizeof(char));
-printf("Size of an int: %d bytes\n", (int)sizeof(int));
-printf("Size of a long int: %d bytes\n", (int)sizeof(long int));
-printf("Size of a long long int: %d bytes\n", (int)sizeof(long long int));
-printf("Size of float: %d bytes\n", (int)sizeof(float));
+printf("Size of char: %d bytes\n", sizeof(char));
+printf("Size of an int: %d bytes\n", sizeof(int));
+printf("Size of a long int: %d bytes\n", sizeof(long int));
+printf("Size of a long long int: %d bytes\n", sizeof(long long int));
+printf("Size of float: %d bytes\n", sizeof(float));
 return (0);
 }
